Update AmbrosialEncrypt.cs (silly edition)
iceayy contribution attempt #1

removed alphaChars var ssince it was declared as private static, but never used outside of the GenerateRandomString method.. so we get rid of it!!! no more!! we hate!!!!!!!!!

removed checked keyword from GenerateRandomString method. it WAS used to make sure that integer arithmetic operations dont overflow, but this simply does not lol because the min value of the int is larger than the length of the alphachars string

replaced the strings.mid method with the char indexy thingamabooby. was uused to extract a single char from a string, buuuut we simplify by using the char indexer so it can be used to access a single char at a specific index in a string.

my c# skills are so silly
make sure to like and subscribe for more silly pull requests on the silliest website half of all programs rely on to not die randomly.